Refactoring Methods & Edit Profile Boilerplate
