+++ front-end/iOS/Planto/Planto/User/EditProfileViewController.swift
@@ -1,0 +1,33 @@
+//
+//  EditProfileViewController.swift
+//  Planto
+//
+//  Created by Aaron Lee on 8/9/20.
+//  Copyright © 2020 D&D. All rights reserved.
+//
+
+import UIKit
+
+class EditProfileViewController: UIViewController {
+    
+    // MARK: - IBOutlets
+    @IBOutlet weak var txtEmail: UITextField!
+    @IBOutlet weak var txtNickName: UITextField!
+    @IBOutlet weak var txtPassword: UIStackView!
+    @IBOutlet weak var txtNewPassword: UIStackView!
+    @IBOutlet weak var txtVerifyNewPassword: UIStackView!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+    }
+    
+    // MARK: - IBActions
+    
+    @IBAction func btnConfirmDidTap(_ sender: UIButton) {
+        
+    }
+    
+    // MARK: - Methods
+    
+}
