+++ front-end/iOS/Planto/Planto/User/JoinViewController.swift
@@ -28,7 +28,7 @@ class JoinViewController: UIViewController {
 // MARK: - IBActions
 extension JoinViewController {
     @IBAction func btnSubmit(_ sender: UIButton) {
-        let user = User()
+        let user = UserUtill()
         
         if (txtEmail.text == "" || txtNickName.text == "" || txtPassword.text == "" || txtConfirmPassword.text == "") {
             alert(message: "입력하시지 않은 정보가 있는지 확인해주세요")
@@ -52,19 +52,15 @@ extension JoinViewController {
         if (responseCode == 200) {
             isAuthenticated = true
             isAutoLogin = true
-            saveUserInfo()
+            
+            UserUtill().saveAllUserDefaults(authenticatedFlag: isAuthenticated,
+                                            autoLoginFlag: isAutoLogin,
+                                            email: txtEmail.text!,
+                                            password: txtPassword.text!)
             // To Do: Request Log the Created User In
         } else {
             alert(message: "입력하신 정보를 확인해주세요")
         }
-    }
-    
-    func saveUserInfo() {
-        UserDefaults.standard.set(isAuthenticated, forKey: Constants.User.Info.Authenticated.rawValue)
-        UserDefaults.standard.set(isAutoLogin, forKey: Constants.User.Info.AutoLogIn.rawValue)
-        UserDefaults.standard.set(txtEmail.text, forKey: Constants.User.Info.Email.rawValue)
-        UserDefaults.standard.set(txtNickName.text, forKey: Constants.User.Info.NickName.rawValue)
-        UserDefaults.standard.set(txtPassword.text, forKey: Constants.User.Info.Password.rawValue)
     }
     
     func alert(message: String) {
